Added CONSTANTS.ts / ENUMS / url / cmd helper
